Final Project from CISC 3440 Machine Learning Fall 2023
This project uses a simple CNN to classify images of the brain for different tumors. The details of the project are in the 3440Project notebook.

The data is from a somewhat popular kaggle dataset that has separated classes for different and no brain tumors. (https://www.kaggle.com/datasets/masoudnickparvar/brain-tumor-mri-dataset/data)

I worked on it along with two other people. I handled the preprocessing and data augmentation. I also took the code from my partners' individual notebooks and refactored it so that it worked with my preprocessing and data augmentation.